requirements.txt
requests>=2.31.0
google-cloud-storage>=2.10.0
schedule>=1.2.0
python-dotenv>=1.0.0
pandas>=2.0.0